fix: 댓글 작성 시 댓글 개수를 DB에 저장 하는 방식을 redis를 사용하는 방식으로 변경

+++ src/main/java/com/hong/ForPaw/controller/DTO/PostResponse.java
@@ -22,7 +22,7 @@ public class PostResponse {
                           String title,
                           String content,
                           LocalDateTime date,
-                          Integer commentNum,
+                          Long commentNum,
                           Integer likeNum,
                           String imageURL){}
 
@@ -39,7 +39,7 @@ public class PostResponse {
                                   String title,
                                   String content,
                                   LocalDateTime date,
-                                  Integer commentNum,
+                                  Long commentNum,
                                   Integer likeNum,
                                   List<PostImageDTO> images,
                                   List<CommentDTO> comments){}
